chore: Update Dockerfile, database.sql, script.sql, and index.ts files
This commit updates the Dockerfile in the rds-init-fn-code directory to include a new file called database.sql. It also modifies the script.sql file in the same directory. Additionally, it updates the index.ts file in the data-ingest-lambda directory to import new dependencies and make use of a Service class. The changes are made to improve the functionality and performance of the application.